% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_skew_symmetric_expm.R
\name{expm}
\alias{expm}
\title{Computes the Exponential of a 3D Skew Symmetric Matrix.}
\usage{
expm(skew_symmetric_matrix)
}
\arguments{
\item{skew_symmetric_matrix}{A 3-by-3 skew-symmetric matrix.}
}
\value{
A 3-by-3 rotation matrix representing the exponential 
of the specified skew-symmetric matrix.
}
\description{
The exponential of a skew-symmetric matrix is computed by means
of the Rodrigues' formula.
}
