% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_rotation_logm.R
\name{logm}
\alias{logm}
\title{Computes the Logarithm of a 3D Rotation Matrix.}
\usage{
logm(rotation_matrix)
}
\arguments{
\item{rotation_matrix}{A 3-by-3 rotation matrix.}
}
\value{
A 3-by-3 skew-symmetric matrix representing the logarithm 
of the specified rotation matrix.
}
\description{
Computes the Logarithm of a 3D Rotation Matrix.
}
